Fix Go 1.22 programs
Go 1.22 added new checks on startup where it checks if the standard fds
are accessible. It does that by using fctnl syscall. This means that
this syscall may not return an error.

Also Go did some page allocations using virtual addresses where the
pages are 1:1 mapped for the kernel. This caused some mapping conflicts.
As mmap is not required to honor the requested start_addr and is only
used as a hint, the kernel can set the start address to be higher than
the highest physical address of the pages to prevent this collision.