Fix bug where arrays are copied to stack
When using range on an array it causes go to copy the array on to the
stack and iterate over that. With sufficiently big arrays this causes a
stack overflow, which causes very weird behaviours, like page tables
suddently containing invalid entries. Using a slice prevents this, as
the slice fundamentally works as a pointer to the array.