Refactor kernel to be in a package called kernel and rename go module
This is the start of refactoring the kernel to put individual components
in their own packages. For this refactor to work there has to be added a
main package with a main-function-stub. This is needed such that the go
compiler emits an ELF file. Otherwise the go compiler would emit a
go-archive which is not understood by other tools to link the final
kernel image.