Move prints to its own package
In order to prevent cyclic imports the print code no longer references
text_mode or serial writers directly. However, to keep printing
convenient the state is needed for the default writers. In order to
support to print both to the screen and to the serial device at the same
time multiple writers can be registered for log, error and debug.